3/20

+++ src/main/java/knou/lms/srvy/vo/SrvyPtcpVO.java
@@ -1,5 +1,7 @@
 package knou.lms.srvy.vo;
 
+import java.math.BigDecimal;
+
 import knou.lms.common.vo.DefaultVO;
 
 public class SrvyPtcpVO extends DefaultVO {
@@ -7,14 +9,14 @@ public class SrvyPtcpVO extends DefaultVO {
 	private static final long serialVersionUID = 2656613132560194056L;
 
 	// TB_LMS_SRVY_PTCP ( 설문참여 )
-	private String  srvyPtcpId;		// 설문참여아이디
-	private String  srvyId;			// 설문아이디
-	private String  ptcpDttm;		// 참여일시
-	private String  cntnDvcTycd;	// 접속기기유형코드
-	private String  cntnip;			// 접속아이피
-	private Integer ptcpEvlScr;		// 참여평가점수
-	private String  profMemo;		// 교수메모
-	private String  srvyPtcpEvlyn;	// 설문참여평가여부
+	private String  	srvyPtcpId;		// 설문참여아이디
+	private String  	srvyId;			// 설문아이디
+	private String  	ptcpDttm;		// 참여일시
+	private String  	cntnDvcTycd;	// 접속기기유형코드
+	private String  	cntnip;			// 접속아이피
+	private BigDecimal 	ptcpEvlScr;		// 참여평가점수
+	private String  	profMemo;		// 교수메모
+	private String  	srvyPtcpEvlyn;	// 설문참여평가여부
 
 	public String getSrvyPtcpId() {
 		return srvyPtcpId;
@@ -31,7 +33,7 @@ public class SrvyPtcpVO extends DefaultVO {
 	public String getCntnip() {
 		return cntnip;
 	}
-	public Integer getPtcpEvlScr() {
+	public BigDecimal getPtcpEvlScr() {
 		return ptcpEvlScr;
 	}
 	public String getProfMemo() {
@@ -55,7 +57,7 @@ public class SrvyPtcpVO extends DefaultVO {
 	public void setCntnip(String cntnip) {
 		this.cntnip = cntnip;
 	}
-	public void setPtcpEvlScr(Integer ptcpEvlScr) {
+	public void setPtcpEvlScr(BigDecimal ptcpEvlScr) {
 		this.ptcpEvlScr = ptcpEvlScr;
 	}
 	public void setProfMemo(String profMemo) {
